+++ src/lib.rs
@@ -1,5 +1,6 @@
 pub mod faucet;
 pub mod grpc_client;
+pub mod mq;
 pub mod pb;
 pub mod storage;
 pub mod tele_out;
