+++ src/mq/mod.rs
@@ -1,3 +1,2 @@
 pub mod messages;
 pub mod msg_consumer;
-pub mod msg_loader;
